Refactor: split of pauli_tracker wrapper
The only intersection between mbqc_scheduling and pauli_tracker is the
PartialOrderGraph type. Because of pyo3 #1444 I extended pauli_tracker.
It worked relatively good, but since the intersection is so small the
additional maintaining overhead is not worth add (vs an additional
python to rust transformation) and it is also cleaner to have this
packages separated for publishing.